fix: keep curl quiet and read the block size without a lossy cast
Nothing drains curl's stderr until the transfer is over, so a progress meter
writing for an hour would fill the pipe and stop the download dead.
`--silent --show-error` leaves only the one line a failure needs, which is what
the `download_failed` envelope carries.

`statvfs` reports `f_frsize` as a `c_ulong`, so it is read as one rather than
through a `u32` that would truncate an unusual block size.

+++ cli/src/model/disk.rs
@@ -48,7 +48,7 @@ fn statvfs_available(path: &Path) -> Option<u64> {
         stats.assume_init()
     };
 
-    let block = u64::from(stats.f_frsize as u32).max(1);
+    let block = (stats.f_frsize as u64).max(1);
     Some(block.saturating_mul(stats.f_bavail as u64))
 }
